Move isArray into src/core/types/array
