+++ build-system/test-configs/dep-check-config.js
@@ -87,6 +87,7 @@ exports.rules = [
       '3p/**->src/utils/object.js',
       '3p/**->src/utils/promise.js',
       '3p/**->src/log.js',
+      '3p/**->src/core/types/array.js',
       '3p/**->src/types.js',
       '3p/**->src/string.js',
       '3p/**->src/style.js',
@@ -428,6 +429,7 @@ exports.rules = [
     allowlist: [
       'src/polyfills/fetch.js->src/log.js',
       'src/polyfills/fetch.js->src/types.js',
+      'src/polyfills/fetch.js->src/core/types/array.js',
       'src/polyfills/fetch.js->src/json.js',
       'src/polyfills/fetch.js->src/utils/object.js',
       'src/polyfills/fetch.js->src/utils/bytes.js',
