+++ extensions/amp-a4a/0.1/amp-a4a.js
@@ -72,7 +72,8 @@ import {
   measureIntersection,
 } from '../../../src/utils/intersection';
 import {isAdPositionAllowed} from '../../../src/ad-helper';
-import {isArray, isEnumValue, isObject} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
+import {isEnumValue, isObject} from '../../../src/types';
 import {listenOnce} from '../../../src/event-helper';
 import {
   observeWithSharedInOb,
