+++ extensions/amp-a4a/0.1/amp-ad-utils.js
@@ -15,7 +15,8 @@
  */
 import {Services} from '../../../src/services';
 import {dev} from '../../../src/log';
-import {isArray, isObject} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
+import {isObject} from '../../../src/types';
 import {isSecureUrlDeprecated} from '../../../src/url';
 import {parseExtensionUrl} from '../../../src/service/extension-script';
 import {parseJson} from '../../../src/json';
