+++ extensions/amp-access/0.1/amp-access-iframe.js
@@ -21,7 +21,7 @@ import {assertHttpsUrl, parseUrlDeprecated} from '../../../src/url';
 import {dev, userAssert} from '../../../src/log';
 import {dict} from '../../../src/utils/object';
 import {getMode} from '../../../src/mode';
-import {isArray} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
 import {parseJson} from '../../../src/json';
 import {toggle} from '../../../src/style';
