+++ extensions/amp-access/0.1/amp-access.js
@@ -29,7 +29,7 @@ import {dict} from '../../../src/utils/object';
 import {getSourceOrigin} from '../../../src/url';
 import {getValueForExpr, tryParseJson} from '../../../src/json';
 import {installStylesForDoc} from '../../../src/style-installer';
-import {isArray} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
 import {isJsonScriptTag} from '../../../src/dom';
 import {listenOnce} from '../../../src/event-helper';
 import {triggerAnalyticsEvent} from '../../../src/analytics';
