+++ extensions/amp-animation/0.1/test/test-web-animations.js
@@ -19,7 +19,8 @@ import {Services} from '../../../../src/services';
 import {WebAnimationPlayState} from '../web-animation-types';
 import {closestAncestorElementBySelector} from '../../../../src/dom';
 import {htmlFor, htmlRefs} from '../../../../src/static-template';
-import {isArray, isObject} from '../../../../src/types';
+import {isArray} from '../../../../src/core/types/array';
+import {isObject} from '../../../../src/types';
 import {poll} from '../../../../testing/iframe';
 import {user} from '../../../../src/log';
