+++ extensions/amp-auto-lightbox/0.1/test/test-amp-auto-lightbox.js
@@ -36,7 +36,7 @@ import {Services} from '../../../../src/services';
 import {Signals} from '../../../../src/utils/signals';
 import {createElementWithAttributes} from '../../../../src/dom';
 import {htmlFor} from '../../../../src/static-template';
-import {isArray} from '../../../../src/types';
+import {isArray} from '../../../../src/core/types/array';
 import {tryResolve} from '../../../../src/utils/promise';
 
 const TAG = 'amp-auto-lightbox';
