+++ extensions/amp-autocomplete/0.1/amp-autocomplete.js
@@ -35,7 +35,8 @@ import {dict, hasOwn, map, ownProperty} from '../../../src/utils/object';
 import {getValueForExpr, tryParseJson} from '../../../src/json';
 import {includes} from '../../../src/string';
 import {isAmp4Email} from '../../../src/format';
-import {isArray, isEnumValue} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
+import {isEnumValue} from '../../../src/types';
 import {mod} from '../../../src/utils/math';
 import {once} from '../../../src/utils/function';
 import {removeChildren, tryFocus} from '../../../src/dom';
