+++ extensions/amp-bind/0.1/bind-impl.js
@@ -33,11 +33,17 @@ import {deepEquals, getValueForExpr, parseJson} from '../../../src/json';
 import {deepMerge, dict, map} from '../../../src/utils/object';
 import {dev, devAssert, user} from '../../../src/log';
 import {escapeCssSelectorIdent} from '../../../src/css';
-import {findIndex, remove, toArray} from '../../../src/core/types/array';
+import {
+  findIndex,
+  isArray,
+  remove,
+  toArray,
+} from '../../../src/core/types/array';
 import {getMode} from '../../../src/mode';
 import {invokeWebWorker} from '../../../src/web-worker/amp-worker';
 import {isAmp4Email} from '../../../src/format';
-import {isArray, isFiniteNumber, isObject} from '../../../src/types';
+
+import {isFiniteNumber, isObject} from '../../../src/types';
 import {reportError} from '../../../src/error';
 import {rewriteAttributesForElement} from '../../../src/url-rewrite';
