+++ extensions/amp-experiment/1.0/variant.js
@@ -18,7 +18,8 @@ import {Deferred} from '../../../src/utils/promise';
 import {Services} from '../../../src/services';
 import {dev, userAssert} from '../../../src/log';
 import {hasOwn} from '../../../src/utils/object';
-import {isArray, isObject} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
+import {isObject} from '../../../src/types';
 
 export const ATTR_PREFIX = 'amp-x-';
 const nameValidator = /^[\w-]+$/;
