+++ extensions/amp-form/0.1/amp-form.js
@@ -64,13 +64,14 @@ import {getMode} from '../../../src/mode';
 import {installFormProxy} from './form-proxy';
 import {installStylesForDoc} from '../../../src/style-installer';
 import {isAmp4Email} from '../../../src/format';
-import {isArray, toWin} from '../../../src/types';
+import {isArray, toArray} from '../../../src/core/types/array';
 import {
   setupAMPCors,
   setupInit,
   setupInput,
 } from '../../../src/utils/xhr-utils';
-import {toArray} from '../../../src/core/types/array';
+
+import {toWin} from '../../../src/types';
 import {triggerAnalyticsEvent} from '../../../src/analytics';
 import {tryParseJson} from '../../../src/json';
