+++ extensions/amp-list/0.1/amp-list.js
@@ -53,7 +53,7 @@ import {getMode} from '../../../src/mode';
 import {getSourceOrigin} from '../../../src/url';
 import {getValueForExpr} from '../../../src/json';
 import {isAmp4Email} from '../../../src/format';
-import {isArray} from '../../../src/types';
+import {isArray, toArray} from '../../../src/core/types/array';
 import {isExperimentOn} from '../../../src/experiments';
 import {px, setImportantStyles, setStyles, toggle} from '../../../src/style';
 import {setDOM} from '../../../third_party/set-dom/set-dom';
@@ -62,7 +62,6 @@ import {
   setupInput,
   setupJsonFetchInit,
 } from '../../../src/utils/xhr-utils';
-import {toArray} from '../../../src/core/types/array';
 
 /** @const {string} */
 const TAG = 'amp-list';
