+++ extensions/amp-next-page/0.1/config.js
@@ -20,7 +20,7 @@ import {
   getSourceUrl,
   resolveRelativeUrl,
 } from '../../../src/url';
-import {isArray} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
 import {user, userAssert} from '../../../src/log';
 
 const ADSENSE_REC_ORIGIN = 'https://googleads.g.doubleclick.net';
