+++ extensions/amp-story/1.0/amp-story-access.js
@@ -24,7 +24,8 @@ import {closest, copyChildren, removeChildren} from '../../../src/dom';
 import {dev, user} from '../../../src/log';
 import {getStoryAttributeSrc} from './utils';
 import {htmlFor} from '../../../src/static-template';
-import {isArray, isObject} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
+import {isObject} from '../../../src/types';
 import {parseJson} from '../../../src/json';
 import {setImportantStyles} from '../../../src/style';
