+++ extensions/amp-subscriptions/0.1/local-subscription-platform-iframe.js
@@ -20,7 +20,7 @@ import {LocalSubscriptionBasePlatform} from './local-subscription-platform-base'
 import {Messenger} from '../../amp-access/0.1/iframe-api/messenger';
 import {assertHttpsUrl, parseUrlDeprecated} from '../../../src/url';
 import {devAssert, userAssert} from '../../../src/log';
-import {isArray} from '../../../src/types';
+import {isArray} from '../../../src/core/types/array';
 import {parseJson} from '../../../src/json';
 import {toggle} from '../../../src/style';
