+++ src/base-element.js
@@ -21,7 +21,8 @@ import {devAssert, user, userAssert} from './log';
 import {dispatchCustomEvent} from './dom';
 import {getData, listen, loadPromise} from './event-helper';
 import {getMode} from './mode';
-import {isArray, toWin} from './types';
+import {isArray} from './core/types/array';
+import {toWin} from './types';
 
 /**
  * Base class for all custom element implementations. Instead of inheriting
