+++ src/multidoc-manager.js
@@ -27,7 +27,8 @@ import {dev, user} from './log';
 import {disposeServicesForDoc, getServicePromiseOrNullForDoc} from './service';
 import {getMode} from './mode';
 import {installStylesForDoc} from './style-installer';
-import {isArray, isObject} from './types';
+import {isArray} from './core/types/array';
+import {isObject} from './types';
 import {parseExtensionUrl} from './service/extension-script';
 import {parseUrlDeprecated} from './url';
 import {setStyle} from './style';
