+++ src/service/owners-impl.js
@@ -18,7 +18,7 @@ import {OwnersInterface} from './owners-interface';
 import {Resource} from './resource';
 import {Services} from '../services';
 import {devAssert} from '../log';
-import {isArray} from '../types';
+import {isArray} from '../core/types/array';
 import {registerServiceBuilderForDoc} from '../service';
 
 /**
