+++ src/utils/xhr-utils.js
@@ -17,7 +17,7 @@
 import {Services} from '../services';
 import {devAssert, user, userAssert} from '../log';
 import {dict, map} from './object';
-import {fromIterator} from '../core/types/array';
+import {fromIterator, isArray} from '../core/types/array';
 import {
   getCorsUrl,
   getWinOrigin,
@@ -26,9 +26,10 @@ import {
   serializeQueryString,
 } from '../url';
 import {getMode} from '../mode';
-import {isArray, isObject} from '../types';
+
 import {isExperimentOn} from '../experiments';
 import {isFormDataWrapper} from '../form-data-wrapper';
+import {isObject} from '../types';
 
 /** @private @const {!Array<string>} */
 const allowedMethods_ = ['GET', 'POST'];
